requirements.txt
# Python Libraries 

PyGithub==1.59.0
langchain==0.1.16
langchain-core==0.1.42
openai==1.57.2
tiktoken==0.8.0
streamlit==1.41.0
python-dotenv==1.0.1
langchain-google-genai==1.0.1